+++ src/socialsim4/core/actions/policy_feedback_actions.py
@@ -4,6 +4,9 @@ from socialsim4.core.action import Action
 class ReportUpwardAction(Action):
     NAME = "report_upward"
     DESC = "Privately report execution difficulty to a direct superior."
+    REPROMPT_PARAM = "message"
+    REPROMPT_SCENE_TYPES = {"policy_cascade_scene"}
+    REPROMPT_TASK_MODES = {"follow_up", "follow_up_thread"}
     INSTRUCTION = """- report_upward: Report execution difficulty to a direct superior
   <Action name=\"report_upward\"><target>Name</target><message>Your difficulty report</message></Action>
 """
@@ -15,6 +18,9 @@ class ReportUpwardAction(Action):
 class EscalateComplaintAction(Action):
     NAME = "escalate_complaint"
     DESC = "Escalate a complaint or risk report to a higher-level superior."
+    REPROMPT_PARAM = "message"
+    REPROMPT_SCENE_TYPES = {"policy_cascade_scene"}
+    REPROMPT_TASK_MODES = {"follow_up", "follow_up_thread"}
     INSTRUCTION = """- escalate_complaint: Escalate a complaint to a higher-level superior
   <Action name=\"escalate_complaint\"><target>Name</target><message>Your escalation report</message></Action>
 """
@@ -26,6 +32,9 @@ class EscalateComplaintAction(Action):
 class ConsultPeerAction(Action):
     NAME = "consult_peer"
     DESC = "Privately consult or coordinate with a connected peer on the same tier."
+    REPROMPT_PARAM = "message"
+    REPROMPT_SCENE_TYPES = {"policy_cascade_scene"}
+    REPROMPT_TASK_MODES = {"follow_up", "follow_up_thread"}
     INSTRUCTION = """- consult_peer: Privately consult a connected peer on the same tier
   <Action name=\"consult_peer\"><target>Name</target><message>Your consultation message</message></Action>
 """
@@ -37,6 +46,9 @@ class ConsultPeerAction(Action):
 class NotifySubordinateAction(Action):
     NAME = "notify_subordinate"
     DESC = "Privately notify a directly connected subordinate."
+    REPROMPT_PARAM = "message"
+    REPROMPT_SCENE_TYPES = {"policy_cascade_scene"}
+    REPROMPT_TASK_MODES = {"follow_up", "follow_up_thread"}
     INSTRUCTION = """- notify_subordinate: Privately notify a directly connected subordinate
   <Action name=\"notify_subordinate\"><target>Name</target><message>Your notification</message></Action>
 """
@@ -48,6 +60,9 @@ class NotifySubordinateAction(Action):
 class AnnouncePolicyAdjustmentAction(Action):
     NAME = "announce_policy_adjustment"
     DESC = "Issue a new policy adjustment announcement that reopens cascade transmission."
+    REPROMPT_PARAM = "message"
+    REPROMPT_SCENE_TYPES = {"policy_cascade_scene"}
+    REPROMPT_TASK_MODES = {"follow_up", "follow_up_thread"}
     INSTRUCTION = """- announce_policy_adjustment: Issue a policy adjustment announcement
   <Action name=\"announce_policy_adjustment\"><message>Your adjustment announcement</message></Action>
 """
