+++ src/socialsim4/core/agent/agent.py
@@ -561,6 +561,98 @@ Use the above context to inform your responses when relevant.
         if not success:
             return {}
 
+        # --- Reprompt handling for actions requiring free-text input ---
+        reprompt_storage_handled = False
+        action_lookup = {
+            getattr(action, "NAME", ""): action
+            for action in self.action_space
+        }
+
+        for item in action_data:
+            action_payload = item.get("action") or {}
+            if type(action_payload) is dict:
+                action_name = str(action_payload.get("name") or action_payload.get("action") or "").strip()
+            else:
+                action_name = str(action_payload or item.get("action") or item.get("name") or "").strip()
+
+            if not action_name:
+                continue
+
+            action_def = action_lookup.get(action_name)
+            if not action_def:
+                continue
+
+            reprompt_param = getattr(action_def, "REPROMPT_PARAM", None)
+            if not reprompt_param:
+                continue
+
+            reprompt_scene_types = getattr(action_def, "REPROMPT_SCENE_TYPES", None)
+            if reprompt_scene_types:
+                scene_type = getattr(scene, "TYPE", "") if scene else ""
+                if scene_type not in reprompt_scene_types:
+                    continue
+
+            reprompt_task_modes = getattr(action_def, "REPROMPT_TASK_MODES", None)
+            if reprompt_task_modes:
+                task_mode = scene._effective_task_mode_for(self)
+                if task_mode not in reprompt_task_modes:
+                    continue
+
+            existing_value = ""
+            if type(action_payload) is dict:
+                existing_value = str(action_payload.get(reprompt_param) or "").strip()
+            else:
+                existing_value = str(item.get(reprompt_param) or "").strip()
+            if existing_value:
+                continue
+
+            self.short_memory.append("assistant", llm_output)
+            if self.log_event:
+                self.log_event(
+                    "agent_ctx_delta",
+                    {"agent": self.name, "role": "assistant", "content": llm_output},
+                )
+            reprompt_storage_handled = True
+
+            reprompt_instruction = (
+                f"You selected the '{action_name}' action. "
+                f"Now write your {reprompt_param} (plain text only, no JSON):"
+            )
+            self.short_memory.append("user", reprompt_instruction)
+            if self.log_event:
+                self.log_event(
+                    "agent_ctx_delta",
+                    {"agent": self.name, "role": "user", "content": reprompt_instruction},
+                )
+
+            reprompt_ctx = self.short_memory.searilize(dialect="default")
+            reprompt_ctx.insert(0, {"role": "system", "content": system_prompt})
+
+            reprompt_output = self.call_llm(clients, reprompt_ctx).strip()
+
+            if type(action_payload) is dict:
+                action_payload[reprompt_param] = reprompt_output
+            else:
+                item[reprompt_param] = reprompt_output
+
+            self.short_memory.append("assistant", reprompt_output)
+            if self.log_event:
+                self.log_event(
+                    "agent_ctx_delta",
+                    {"agent": self.name, "role": "assistant", "content": reprompt_output},
+                )
+
+            try:
+                with open(_debug_file, 'a', encoding='utf-8') as f:
+                    f.write(f"\n--- REPROMPT for '{action_name}' ---\n")
+                    f.write(f"Instruction: {reprompt_instruction}\n")
+                    f.write(f"Response: {reprompt_output}\n")
+                    f.write(f"--- END REPROMPT ---\n\n")
+            except Exception:
+                pass
+
+        # --- End reprompt handling ---
+
         # Store a compact assistant memory instead of raw JSON to reduce self-copying
         memory_parts = []
         scene_type = getattr(scene, "TYPE", "") if scene else ""
@@ -601,7 +693,7 @@ Use the above context to inform your responses when relevant.
                     fallback_parts.append(f"[Action] {action_name}")
             assistant_memory = "\n".join(fallback_parts).strip()
 
-        if assistant_memory:
+        if assistant_memory and not reprompt_storage_handled:
             self.short_memory.append("assistant", assistant_memory)
             if self.log_event:
                 self.log_event(
