+++ src/socialsim4/core/scenes/policy_cascade/__init__.py
@@ -1,0 +1,3 @@
+from .scene import PolicyCascadeScene
+
+__all__ = ["PolicyCascadeScene"]
